+++ go.mod
@@ -1,0 +1,3 @@
+module premestitev.sdl/v2
+
+go 1.23.5
